bridge/requirements.txt
influxdb_client
redis
uvicorn
dotenv
fastapi
pydantic
pyserial
